Move xvc_session fixture to root conftest to fix import collision
The xilinx conftest.py was adding its directory to sys.path[0],
which caused pytest to resolve 'from conftest import wait_for'
to the wrong conftest module during full test collection.
